update git erros

+++ production/solver/cuda1/src/lib/json.cpp
@@ -15,7 +15,7 @@ class JSON {
    void append(string const& name, double const& scalar);
    void append(string const& name, float const& scalar);
    void append(string const& name, int const& scalar);
-   void append(string const& name, long long int const& scalar);
+   void append(string const& name, long int const& scalar);
 
    string dump();
  private:
@@ -147,7 +147,7 @@ void JSON::append(string const& name, int const& scalar) {
    this->output += "\"" + name + "\":" + to_string(scalar);
 };
 
-void JSON::append(string const& name, long long int const& scalar) {
+void JSON::append(string const& name, long int const& scalar) {
    if (this->output.size() > 1) {
       this->output += ",";
    }
